UI modification & Code refactor
- Change the primary color to light black
- Remove default application background style
- Remove movie list recycler view margins
- Organise classes in the appropriate packages
