+++ app/src/main/java/com/oganbelema/popularmovies/movie/viewmodel/MovieViewModelFactory.java
@@ -1,11 +1,10 @@
-package com.oganbelema.popularmovies.movie;
+package com.oganbelema.popularmovies.movie.viewmodel;
 
 import androidx.annotation.NonNull;
 import androidx.lifecycle.ViewModel;
 import androidx.lifecycle.ViewModelProvider;
 
-import com.oganbelema.popularmovies.movie.MovieRepository;
-import com.oganbelema.popularmovies.movie.MovieViewModel;
+import com.oganbelema.popularmovies.movie.repository.MovieRepository;
 
 import javax.inject.Inject;
 import javax.inject.Singleton;
