+++ app/src/main/res/layout/activity_movie_detail.xml
@@ -4,7 +4,8 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".movie.moviedetail.MovieDetailActivity">
+    android:background="@color/colorPrimary"
+    tools:context=".movie.ui.MovieDetailActivity">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
