+++ app/src/main/res/layout/activity_popular_movies_list.xml
@@ -4,7 +4,8 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".movie.movielist.MovieListActivity">
+    android:background="@color/colorPrimary"
+    tools:context=".movie.ui.MovieListActivity">
 
     <androidx.constraintlayout.widget.Group
         android:id="@+id/loaderViews"
@@ -44,10 +45,6 @@
         android:id="@+id/moviesRecyclerView"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:layout_marginStart="8dp"
-        android:layout_marginTop="8dp"
-        android:layout_marginEnd="8dp"
-        android:layout_marginBottom="8dp"
         android:visibility="gone"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
