+++ app/src/main/res/values/colors.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <color name="colorPrimary">#696969</color>
+    <color name="colorPrimary">#191919</color>
     <color name="colorPrimaryDark">#000000</color>
     <color name="colorAccent">#dd1c1c</color>
 </resources>
